remove dependence on oce and ocedata
This dependence will create a circularity that will make it impossible
to submit gsw to CRAN, since oce will soon depend on gsw. The main thing
this means is that the vignette cannot use oce to create figures.
Therefore, I've altered the vignette code (as documented in the .Rnw
file), and added the relevant PNG and PDF files to the repo.