improve wrappers
I was lazy before, and had not written macros for wrappers for the gsw C
functions that yield multiple scalars. (There is no point in writing
macros for C functions that return vectors, since each is distinct
enough that we would need almost one macro per function.) This shortens
src/wrappers.c by about 20 percent, but the more important reason for
doing this is that it should reduce coding errors in the 40 or so
functions that are still left to wrap. (I estimate 1 coding error per 5
functions, and maybe as high as 4 per 5, at the start of a coding
session.)