NA handling tests (issue 37)
I realized that the code prior to yesterday should have handled the case
in which some argument values are NA. So I removed yesterday's change
from gsw_CT_from_t() as a test. The results are confusing, because I
cannot reproduce the bug noted in
https://github.com/TEOS-10/GSW-R/issues/37. My test was on my work
machine. This afternoon, I'll try on my home machine. If I cannot
reproduce the bug, I'll revert the subset of yesterday's changes that
relate to this.  (It's not that I think those changes will hurt
anything, but if they are not needed, they will just be an instance of
never-called code that will confuse readers.