improved docs wrt ranges; correct commit hash
1. The template for parameter SA now gives a valid range.
2. All density-related functions now have a see-also link to
   gsw_infunnel().
3. The hash listed in the Details section of functions had not been
   updated to the correct value.  roxygen2 templates cannot be used in
the DESCRIPTION file, so I have to remember to update the hash code in
multiple places, and I forgot.  I've never had a report of an incorrect
hash code, but I'd be willing to bet that I make this error in quite a
few releases. Still, I think it's good to refer to the hash, if only to
remind users that we GSW-R is based on GSW-C, which in turn is based on
GSW-Matlab and TEOS-10.